# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Pedeve App Backend API dengan struktur Clean Architecture",
        "title": "Pedeve App Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/document-types": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar semua jenis dokumen yang aktif. User reguler hanya melihat jenis dokumen aktif, superadmin/administrator bisa melihat semua termasuk yang tidak aktif.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Ambil Semua Document Types",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Include inactive document types (hanya superadmin/administrator)",
                        "name": "include_inactive",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar jenis dokumen berhasil diambil",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.DocumentTypeModel"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat jenis dokumen baru. Hanya superadmin dan administrator yang dapat membuat jenis dokumen baru. Jika jenis dokumen dengan nama yang sama sudah ada (tidak aktif), akan diaktifkan kembali.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Buat Document Type Baru",
                "parameters": [
                    {
                        "description": "Document type data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Jenis dokumen berhasil dibuat",
                        "schema": {
                            "$ref": "#/definitions/domain.DocumentTypeModel"
                        }
                    },
                    "400": {
                        "description": "Invalid request atau jenis dokumen sudah ada",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (hanya superadmin/administrator)",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/document-types/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengupdate jenis dokumen (nama atau status aktif/tidak aktif). Hanya superadmin dan administrator yang dapat mengupdate jenis dokumen.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Update Document Type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Document type data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jenis dokumen berhasil diupdate",
                        "schema": {
                            "$ref": "#/definitions/domain.DocumentTypeModel"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (hanya superadmin/administrator)",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Document type tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus jenis dokumen (soft delete: set is_active = false). Hanya superadmin dan administrator yang dapat menghapus. Jika jenis dokumen sudah digunakan oleh dokumen, akan di-soft delete (tidak aktif) agar dokumen yang sudah ada tidak error.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Hapus Document Type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jenis dokumen berhasil dihapus",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (hanya superadmin/administrator)",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Document type tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.DocumentTypeModel": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "description": "User yang membuat",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "description": "Soft delete: false jika dihapus",
                    "type": "boolean"
                },
                "name": {
                    "description": "Nama jenis dokumen (unique)",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "usage_count": {
                    "description": "Jumlah dokumen yang menggunakan jenis ini",
                    "type": "integer"
                }
            }
        },
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token. Atau menggunakan httpOnly cookie (auth_token) untuk keamanan yang lebih baik.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}